Use list instead of get_all_keys
The get_all_keys method requires pagination handling to
retrieve all of the contents of a bucket. The higher-level
list method will return a proper iterator with all of
the items.

Fixes a bug where doing a restore would not actually
return the most recent backup if there were more than
1000 entries in the bucket.